+++ workspaces/core/src/button/private/utils/buttonStyles/getButtonCommonStyles.ts
@@ -6,8 +6,8 @@ const getButtonCommonStyles = (
 ): string =>
   mergeClassnames(
     'relative z-0 flex justify-center items-center font-medium no-underline overflow-hidden',
-    'whitespace-nowrap select-none transition duration-200 active:scale-90',
-    disabled && 'opacity-30 cursor-not-allowed'
+    'whitespace-nowrap select-none transition duration-200',
+    disabled ? 'opacity-30 cursor-not-allowed' : 'active:scale-90'
   );
 
 export default getButtonCommonStyles;